services/inference-core/requirements.txt
fastapi
uvicorn
httpx
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-google-genai==2.0.8
pydantic
pydantic-settings
psycopg2-binary
python-dotenv
